+++ xarray_enmap/cli.py
@@ -10,6 +10,10 @@ import pathlib
 import shutil
 import sys
 import tempfile
+from collections.abc import Iterable
+
+import xarray
+
 from . import xarray_enmap
 
 LOGGER = logging.getLogger(__name__)
@@ -55,6 +59,13 @@ def main():
         help="Higher Zarr output compression. ~25%% smaller than default compression. "
         "Compression process (but not decompression) is much slower.",
     )
+    parser.add_argument(
+        "--datatree",
+        "-d",
+        action="store_true",
+        help="Whether to write the data as datatree. This parameter is only considered when "
+        "the parameter zarr-output is given.",
+    )
     parser.add_argument("--verbose", "-v", action="count", default=0)
     args = parser.parse_args()
     scale_reflectance = not args.raw_reflectance
@@ -79,6 +90,7 @@ def main():
                 args.tiff_output,
                 temp_dir,
                 args.compress,
+                args.datatree,
                 scale_reflectance,
             )
     else:
@@ -87,10 +99,11 @@ def main():
         os.mkdir(temp_dir)
         process(
             args.input_filename,
-            os.path.expanduser(args.output_dir),
+            args.zarr_output,
+            args.tiff_output,
             temp_dir,
             args.compress,
-            args.extract_only,
+            args.datatree,
             scale_reflectance,
         )
 
@@ -101,6 +114,7 @@ def process(
     output_dir_tiff: str,
     temp_dir: str,
     compress: bool = False,
+    open_as_datatree: bool = False,
     scale_reflectance: bool = True,
 ):
     if output_dir_zarr is output_dir_tiff is None:
@@ -130,13 +144,20 @@ def process(
         raise ValueError(
             f"{input_filename} is neither a file nor a directory."
         )
-    for data_dir in data_dirs:
-        if output_dir_tiff is not None:
-            shutil.copytree(
-                data_dir, pathlib.Path(output_dir_tiff) / data_dir.name
-            )
-        if output_dir_zarr is not None:
-            write_zarr(data_dir, output_dir_zarr, compress, scale_reflectance)
+    if output_dir_zarr is not None and open_as_datatree:
+        write_datatree_as_zarr(
+            input_path, data_dirs, output_dir_zarr, compress, scale_reflectance
+        )
+    else:
+        for data_dir in data_dirs:
+            if output_dir_tiff is not None:
+                shutil.copytree(
+                    data_dir, pathlib.Path(output_dir_tiff) / data_dir.name
+                )
+            if output_dir_zarr is not None:
+                write_zarr(
+                    data_dir, output_dir_zarr, compress, scale_reflectance
+                )
 
 
 def write_zarr(
@@ -154,10 +175,43 @@ def write_zarr(
     ds = xarray_enmap.read_dataset_from_inner_directory(
         data_dir, scale_reflectance
     )
-    zarr_args = {
-        "zarr_format": 2,
-        "store": pathlib.Path(output_dir) / (data_dir.name + ".zarr"),
-    }
+    store_path = pathlib.Path(output_dir) / (data_dir.name + ".zarr")
+    zarr_args = _get_zarr_args(compress, store_path)
+    ds.to_zarr(**zarr_args)
+
+
+def write_datatree_as_zarr(
+    input_path: pathlib.Path,
+    data_dirs: Iterable[pathlib.Path | str],
+    output_dir: str,
+    compress: bool = False,
+    scale_reflectance: bool = True,
+):
+    name = input_path.name
+    LOGGER.info(f"Writing {name} to a Zarr archive...")
+    suffixes = input_path.suffixes
+    suffixes.reverse()
+    for suffix in suffixes:
+        name = name.removesuffix(suffix)
+    ensure_module_importable("zarr")
+    LOGGER.info(
+        f"Using {'scaled' if scale_reflectance else 'unscaled'} "
+        f"reflectance."
+    )
+    groups = {}
+    for data_dir in data_dirs:
+        group_name = data_dir if isinstance(data_dir, str) else data_dir.name
+        groups[group_name] = xarray_enmap.read_dataset_from_inner_directory(
+            data_dir, scale_reflectance
+        )
+    dt = xarray.DataTree.from_dict(groups)
+    store_path = pathlib.Path(output_dir) / (name + ".zarr")
+    zarr_args = _get_zarr_args(compress, store_path)
+    dt.to_zarr(**zarr_args)
+
+
+def _get_zarr_args(compress: bool, store_path: str):
+    zarr_args = {"zarr_format": 2, "store": store_path}
     if compress:
         ensure_module_importable("numcodecs")
         import numcodecs
@@ -169,7 +223,7 @@ def write_zarr(
                 )
             }
         }
-    ds.to_zarr(**zarr_args)
+    return zarr_args
 
 
 def ensure_module_importable(module_name: str):